#8: Fixing elevated tests was too difficult because for some reason, mocks get automatically reset across method body boundaries, so setting up a mock in a constructor to be shared by all test methods in a class breaks the tests, so you would have to copy and paste duplicate initialization logic into each test method. This, combined with JustMock being very slow and brittle to integrate into a build, made me decide to move completely to mocking using MidiDeviceFactory, which is a coverage-excluded facade for DryWetMidi's Input/OutpuDevice.GetInput/OutputDeviceByName static methods, which mercifully were improved to return interfaces. Added last missing HUI tests for 100% coverage.

+++ BehringerXTouchExtender/MidiClient.cs
@@ -45,7 +45,7 @@ internal sealed class HuiMidiClient: MidiClient {
     /// </summary>
     public void BeforeWritingToIlluminatedButton(int trackId) {
         if (_mostRecentOutputTrackId != trackId) {
-            ToDevice?.SendEvent(new ControlChangeEvent((SevenBitNumber) 0x0c, (SevenBitNumber) trackId));
+            ToDevice?.SendEvent(new ControlChangeEvent((SevenBitNumber) 0x0c, (SevenBitNumber) (byte) trackId));
             _mostRecentOutputTrackId = trackId;
         }
     }
